refactor(lessons/channel): refatorado o import do dto de criaçao de canal
