+++ src/modules/lessons/infra/typeorm/repositories/Channels.repository.ts
@@ -1,6 +1,6 @@
 import { getRepository, Repository } from 'typeorm';
 
-import { CreateChannelDto } from '@lessons:dtos/CreateChannelDto';
+import { CreateChannelDto } from '@lessons:dtos/CreateChannel.dto';
 import Channel from '@lessons:entities/Channel';
 import IChannelRepositories from '@lessons:irepos/IChannels.repository';
